feat(astro): add experiemntal image service
