+++ examples/astro/src/pages/service.md
@@ -1,0 +1,5 @@
+# Hi
+
+![Food](https://cdn.shopify.com/static/sample-images/garnished.jpeg)
+
+![Lighthouse](../formentor.jpg#w=100)
