+++ packages/astro/index.ts
@@ -1,7 +1,4 @@
-// Do not write code directly here, instead use the `src` folder!
-// Then, use this file to export everything you want your user to access.
-
-// @ts-ignore
+// @ts-expect-error - Astro imports don't work with TypeScript libs :(
 export { default as Image } from "./src/Image.astro";
-// @ts-ignore
+// @ts-expect-error - Astro imports don't work with TypeScript libs :(
 export { default as Source } from "./src/Source.astro";
